retour PlayReturn

+++ Qwirkle.Core.ComplianceContext/Entities/Player.cs
@@ -8,7 +8,7 @@ namespace Qwirkle.Core.ComplianceContext.Entities
         public int GameId { get; set; }
         public byte GamePosition { get; set; }
         public bool GameTurn { get; set; }
-        public byte Points { get; set; }
+        public int Points { get; set; }
         public List<Tile> Tiles { get; set; }
     }
 }
